feat(core): implement threshold core binary (#562)
* feat(core): implement threshold core binary

* fix(coordinator): purge files in integration tests

* fix(coordinator): correctly generate keys for integration test

* fix(coordinator): clean up after integration test

* chore(coordinator): better cli naming

* refactor(coordinator): better cli interface for gen and server

* fix(connector): compilation issue from key API changes

* fix(docker): missing keys in docker build

+++ coordinator/src/consts.rs
@@ -24,7 +24,7 @@ pub const THRESHOLD: usize = 1;
 pub const SEC_PAR: u64 = 128;
 pub const COMPRESSED: bool = true;
 
-pub const MINIMUM_SESSIONS_PREPROC: u128 = 2;
+pub const MINIMUM_SESSIONS_PREPROC: u16 = 2;
 
 lazy_static! {
     pub static ref TEST_CENTRAL_KEY_ID: RequestId =
